Abandoning this attempt for now.
Rust with wasm just isn't ready for primetime, mostly because async is
so half baked, and the tooling isn't ready.

I wanted to load an STL and render it with webgl. Unfortunately, because
we're in a browser context, loading the STL requires the fetch API which
means we need to use an async future to extract this data.

The gymnastics required to do this mean that I'd have to change the
entire program (running `main` from cargo run with a web target) as main
can't be async.

Then you're into learning how to use the wasm-bindgen tooling directly
and all the documentation uses webpack, which is a red flag for feature
creep to me.

At this point, none of this is even fun any more, rather than solving
problems, I'm researching types of glue.